requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
azure-storage-queue==12.10.0
python-dotenv==1.0.1
azure-storage-blob==12.20.0